# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShiftTrack API",
    "version": "0.1.0",
    "description": "Rutas derivadas de los routers implementados. Contrato de comandos generado de Zod. Los payloads de módulos se validan en sus handlers; este inventario todavía no describe todos sus cuerpos."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/schedules/clone": {
      "post": {
        "operationId": "post__schedules_clone",
        "summary": "POST /schedules/clone",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/tasks/catalog": {
      "post": {
        "operationId": "post__tasks_catalog",
        "summary": "POST /tasks/catalog",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/tasks/assignments": {
      "post": {
        "operationId": "post__tasks_assignments",
        "summary": "POST /tasks/assignments",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/tasks/{id}/revert": {
      "post": {
        "operationId": "post__tasks__id__revert",
        "summary": "POST /tasks/{id}/revert",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/tasks/assignments/{id}/deactivate": {
      "post": {
        "operationId": "post__tasks_assignments__id__deactivate",
        "summary": "POST /tasks/assignments/{id}/deactivate",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/shortages/{id}/resolve": {
      "post": {
        "operationId": "post__shortages__id__resolve",
        "summary": "POST /shortages/{id}/resolve",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/users/{id}/update": {
      "post": {
        "operationId": "post__users__id__update",
        "summary": "POST /users/{id}/update",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/suppliers/{id}/update": {
      "post": {
        "operationId": "post__suppliers__id__update",
        "summary": "POST /suppliers/{id}/update",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/supplier-corrections/prepare": {
      "post": {
        "operationId": "post__supplier_corrections_prepare",
        "summary": "POST /supplier-corrections/prepare",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/state": {
      "get": {
        "operationId": "get__devices_state",
        "summary": "GET /devices/state",
        "security": [
          {
            "deviceTransport": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/devices/pause-ack": {
      "post": {
        "operationId": "post__devices_pause_ack",
        "summary": "POST /devices/pause-ack",
        "security": [
          {
            "deviceTransport": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/supplier-corrections": {
      "post": {
        "operationId": "post__supplier_corrections",
        "summary": "POST /supplier-corrections",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/supplier-corrections/cancel": {
      "post": {
        "operationId": "post__supplier_corrections_cancel",
        "summary": "POST /supplier-corrections/cancel",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/expense-categories": {
      "post": {
        "operationId": "post__expense_categories",
        "summary": "POST /expense-categories",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/alerts/settings": {
      "post": {
        "operationId": "post__alerts_settings",
        "summary": "POST /alerts/settings",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
    "/notifications/{id}/retry": {
      "post": {
        "operationId": "post__notifications__id__retry",
        "summary": "POST /notifications/{id}/retry",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada o datos autorizados"
          },
          "400": {
            "description": "Validación"
          },
          "401": {
            "description": "Identidad requerida"
          },
          "403": {
            "description": "Permiso denegado"
          },
          "409": {
            "description": "Conflicto: conservar captura local"
          },
          "500": {
            "description": "Error del servicio"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token (no custom token)"
      },
      "deviceTransport": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Device TOKEN; solo transporte de pendientes y confirmaciones de equipo"
      }
    },
...
  }
}